feat: show channel details when click 'Channel details' in channel options
